@@ -1,17 +1,19 @@
 package com.haras;
 
-import javax.swing.SwingUtilities;
 import com.haras.view.BaseView;
+import javax.swing.SwingUtilities;
+import javax.swing.UIManager;
 
 public class Main {
     public static void main(String[] args) {
-        // Executar na thread da interface gráfica
-        SwingUtilities.invokeLater(new Runnable() {
-            @Override
-            public void run() {
-                BaseView baseView = new BaseView();
-                baseView.setVisible(true);
+        SwingUtilities.invokeLater(() -> {
+            try {
+                UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
+            } catch (Exception e) {
+                e.printStackTrace();
             }
+            
+            new BaseView().setVisible(true);
         });
     }
 }
